Fixed Animation Times and Crouching
-I noticed animations were lagging when switching between states, turns
out there was a FixedTime checkbox that had all the transitions set to
.25 seconds which is where the lag was coming from.
-I also played around with the friction a bit so it feels slightly
better to play at the moment.
-I changed the jumping AddForce to type Impulse, meaning all the force
for the jump is added at once. This should be the basis for our jump and
then we can add a bit of force to it if we hold the button longer.