Wall Jumping
For now, wall jump is in its own component on the player gameobject.

The intent is for the player to stop dropping when clinging to the wall,
but for some reason that is currently not the case (it looks cool as is,
but i dont know why its like that).

The wall jumping is used best when you jump without pressing in any
direction, and then direct the player (for the current implementation)