Reworking the PlayerController
-Removed Input Handler, all that code is now in the player controller
(logic being that if the player controller "controls" the player, it
should handle the input
-The players state is now set when it receive input, so hitting the jump
button now sets the player to "Jump" state and so on
-This all started when i realised GroundCheck wasn't on the player, and
in implementing it decided to try some stuff
-I'm considering finding an alternative to the PlayerState variable.
Currently the player cannot be both "jumping" and "attacking" since it
can only ever have one state. Perhaps a series of booleans will serve us
better in this regard.
-Probably cleaned up some stuff with turrent behaviour and slimes, can't
remember if that was in this commit or the last one. I should start
sticking to one commit per day, or atleast take notes on what im doing
as i do it.