Enemies and Projectiles
Enemies
-Some refactoring regarding enemies, mainly trying to seperate them into
concise components and reuse as much functionality as possible
-A distinction needs to be made between enemies with health, the ability
to move, bodies, etc. and "hazards" such as the Wizward
-The dragon now has a healthbar, and can die
-When an enemy dies, they leave behind some coins
-Moved the code for enemy damaging into the general enemy script and
removed knockback entirely for now
-Removed Death as a component and moved its logic into the Health
component for now
Projectiles
-Projectiles now get retagged when the Player reflects one so that it is
considered a PlayerAttack, essentially allowing reflected projectiles to
hurt enemies

Player
-fixed an issue where the player would wall jump when standing on the
ground against the wall

NOTE:
-Because of searching for the transform.root when warping, there is
issues with transporting things that aren't the player. If you try to
warp an enemy that is a child of the tilemap, the entire tilemap will be
warped around